+++ packages/tinacms/src/toolkit/fields/plugins/mdx-field-plugin/plate/plugins/create-img-plugin/index.tsx
@@ -1,31 +1,28 @@
 import React from 'react';
 import { ImgEmbed } from './component';
-import { ReactEditor } from 'slate-react';
 import { insertInlineElement } from '../core/common';
 import type { Media } from '../../../../../../core/media';
 import { isImage } from '@toolkit/components/media/utils';
-import {
-  createPluginFactory,
-  normalizeEditor,
-  type PlateEditor,
-  setNodes,
-} from '@udecode/plate-common';
+import { createPlatePlugin, PlateEditor } from '@udecode/plate/react';
 
 export const ELEMENT_IMG = 'img';
 
-const createImgPlugin = createPluginFactory({
+const ImgEmbedComponent = (props) => {
+  const handleChange = (values) => {
+    const path = props.path;
+    props.editor.tf.setNodes(values, { at: path });
+  };
+  return <ImgEmbed {...props} onChange={handleChange} />;
+};
+
+const createImgPlugin = createPlatePlugin({
   key: ELEMENT_IMG,
-  isVoid: true,
-  isInline: true,
-  isElement: true,
-  component: (props) => {
-    const handleChange = (values) => {
-      const path = ReactEditor.findPath(props.editor, props.element);
-      setNodes(props.editor, values, { at: path });
-    };
-    return <ImgEmbed {...props} onChange={handleChange} />;
+  node: {
+    isVoid: true,
+    isInline: true,
+    isElement: true,
   },
-});
+}).withComponent(ImgEmbedComponent);
 
 export const insertImg = (editor: PlateEditor, media: Media) => {
   if (isImage(media.src)) {
@@ -46,7 +43,7 @@ export const insertImg = (editor: PlateEditor, media: Media) => {
   }
 
   // Normalizing the editor after insertion
-  normalizeEditor(editor, { force: true });
+  editor.tf.normalize({ force: true });
 };
 
 export default createImgPlugin;