+++ packages/tinacms/src/toolkit/react-sidebar/components/sidebar.tsx
@@ -269,6 +269,7 @@ const Sidebar = ({
             <div className='fixed left-0 top-0 z-overlay h-full transform'>
               <Nav
                 isLocalMode={cms.api?.tina?.isLocalMode}
+                menuIsOpen
                 toggleMenu={toggleMenu}
                 showCollections={isTinaAdminEnabled}
                 collectionsInfo={collectionsInfo}
@@ -377,6 +378,8 @@ const SidebarHeader = ({ branchingEnabled, isLocalMode }) => {
             <button
               className='p-2 hover:bg-gray-100 transition-colors duration-150 ease-in-out rounded'
               onClick={toggleMenu}
+              title='Open navigation menu'
+              aria-label='Open navigation menu'
             >
               <BiMenu className='h-6 w-auto text-gray-600' />
             </button>
@@ -394,6 +397,8 @@ const SidebarHeader = ({ branchingEnabled, isLocalMode }) => {
             <button
               className='p-2 hover:bg-gray-100 transition-colors duration-150 ease-in-out rounded'
               onClick={toggleSidebarOpen}
+              title='Close sidebar'
+              aria-label='Close sidebar'
             >
               <PiSidebarSimpleLight className='h-6 w-auto text-gray-600' />
             </button>